% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{ar_to_StSp}
\alias{ar_to_StSp}
\title{Convert an AR(k) into a state space system using the companion form.}
\usage{
ar_to_StSp(A, Sigma)
}
\arguments{
\item{A}{s x sk matrix of coefficients as estimated in fit_ar.}

\item{Sigma}{sxs covariance matrix of noise. Default: I(s).}
}
\value{
syst StSp object
}
\description{
Convert an AR(k) into a state space system using the companion form.
}
