% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma2StSp.R
\name{arma2StSp}
\alias{arma2StSp}
\title{Convert an ARMA system to an equivalent state space system.}
\usage{
arma2StSp(arma)
}
\arguments{
\item{arma}{a list of matrices a and b and a variance matrix Sigma.}
}
\value{
syst a \code{StSp} object.
}
\description{
This function converts an ARMA system specified in a list with matrices 'a' and 'b' and innovation variance 'Sigma' to an innovation
state space system (F,H,K) and Sigma.
}
