% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{cal_crit}
\alias{cal_crit}
\title{Calculate the criterion function as a function of the parameter vector}
\usage{
cal_crit(theta, n, s, y, pe = TRUE)
}
\arguments{
\item{theta}{vector of parameters}

\item{n}{number of states}

\item{s}{number of outputs}

\item{y}{matrix of observations}

\item{pe}{Boolean whether to use prediction error formulation. Default is \code{TRUE}.}
}
\value{
negative log likelihood of the model
}
\description{
This function calculates the criterion function with help of the Kalman filter as a function of the parameter vector.
It takes the parameter vector theta, the number of states n, the number of outputs s, the data y and a Boolean pe as input.
It returns the negative log likelihood of the model parameters.
}
