% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_est.R
\name{cal_crit_min_dist}
\alias{cal_crit_min_dist}
\title{Calculate the criterion function for minimum distance estimation as a function of the parameter vector}
\usage{
cal_crit_min_dist(theta, n, s, C, lags = NULL, W = NULL)
}
\arguments{
\item{theta}{vector of parameters}

\item{n}{number of states}

\item{s}{number of outputs}

\item{C}{list of estimated covariances.}

\item{lags}{vector of integers containing the lags contained in C.}

\item{W}{matrix for weighing the measurement errors in C.}
}
\value{
criterion value of squared differences.
}
\description{
This function calculates the criterion function as the squared distance between the directly estimated covariances
and the corresponding covariances as a function of the parameter vector.
It takes the parameter vector theta, the number of states n, the number of outputs s, the data C, the lags contained in C and
a weighting matrix W as input.
It returns the corresponding squared difference.
}
