% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_AR_est.R
\name{cal_crit_min_dist_AR}
\alias{cal_crit_min_dist_AR}
\title{Calculate the criterion function for minimum distance estimation as a function of the parameter vector corresponding
to an AR system}
\usage{
cal_crit_min_dist_AR(theta, k, s, C, lags = NULL, W = NULL)
}
\arguments{
\item{theta}{vector of parameters}

\item{k}{integer; lag length of the AR(k) model}

\item{s}{number of outputs}

\item{C}{list of estimated covariances.}

\item{lags}{vector of integers containing the lags contained in C.}

\item{W}{matrix for weighing the measurement errors in C.}
}
\value{
criterion value of squared differences.
}
\description{
This function calculates the criterion function as the squared distance between the directly estimated covariances
and the corresponding covariances as a function of the parameter vector for an AR system.
It takes the parameter vector theta, the lag length p, the number of outputs s, the data C, the lags contained in C and
a weighting matrix W (to account for heteroscedasticity) as input.
It returns the corresponding squared difference.
}
