% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UC_utils.R
\name{cal_crit_uc}
\alias{cal_crit_uc}
\title{unobserved components model: Calculate criterion function / negative log likelihood}
\usage{
cal_crit_uc(theta, y)
}
\arguments{
\item{theta}{vector of parameters}

\item{y}{matrix of observations}
}
\value{
negative log likelihood
}
\description{
This function calculates the criterion function, i.e. the negative log likelihood, for an unobserved components model given the parameter vector \code{theta} and observed data \code{y}.
}
