% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_CVA}
\alias{cal_CVA}
\title{Canonical Variate Analysis (CVA) subspace estimation}
\usage{
cal_CVA(y, f, p, n)
}
\arguments{
\item{y}{matrix of observed data where rows represent variables and columns represent time points.}

\item{f}{integer representing the future time horizon.}

\item{p}{integer representing the past time horizon.}

\item{n}{integer representing the number of states to estimate (state order). Must be less than or equal to \code{f} and \code{p}.}
}
\value{
syst in \code{StSp} object form.
}
\description{
see Bauer, D.; Buschmeier, R. Asymptotic Properties of Estimators for Seasonally Cointegrated
State Space Models Obtained Using the CVA Subspace Method. Entropy 2021, 23, 436
}
\details{
This function performs Canonical Variate Analysis (CVA) to estimate a state space model.
It calculates the system matrices in canonical form based on the input data and specified parameters.
}
\examples{
set.seed(123)
y <- as.matrix(t(stats::arima.sim(n = 100, model = list(ar = c(0.5, -0.2, 0.3)), sd = 1)))
cal_CVA(y, f=5, p=5, n=1)

}
