% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{cal_kf}
\alias{cal_kf}
\title{Calculate the Kalman filter output}
\usage{
cal_kf(theta, n, s, y, pe = TRUE)
}
\arguments{
\item{theta}{list of parameters}

\item{n}{number of states}

\item{s}{number of outputs}

\item{y}{matrix of observations}

\item{pe}{Boolean whether to use prediction error formulation. Default is \code{TRUE}.}
}
\value{
list of Kalman filter results
}
\description{
This function calculates the Kalman filter output using the parameters theta, the number of states n, the number of outputs s, the data y and a Boolean pe as input.
It returns the Kalman filter output as a list containing the estimated states, the residuals, the innovation variance, the Kalman gain and the negative log likelihood.
}
