% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{check_stability}
\alias{check_stability}
\title{check for stability of an AR system (via the eigenvalues of the companion form)}
\usage{
check_stability(object)
}
\arguments{
\item{object}{ar_model}
}
\value{
stab_cond list containing the entries 'check' Boolean (TRUE, if stable) and 'poles' (eigenvalues of companion matrix)
}
\description{
check for stability of an AR system (via the eigenvalues of the companion form)
}
