% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{check_stability.StSp}
\alias{check_stability.StSp}
\title{check for stability of an AR system (via the eigenvalues of the companion form)}
\usage{
\method{check_stability}{StSp}(object)
}
\arguments{
\item{object}{system, \code{StSp} object}
}
\value{
stab_cond list containing the entries 'check' Boolean (TRUE, if stable) and 'poles' (eigenvalues of companion matrix)
}
\description{
check for stability of an AR system (via the eigenvalues of the companion form)
}
