% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{check_stability_ar}
\alias{check_stability_ar}
\title{check for stability of an AR system (via the eigenvalues of the companion form)}
\usage{
check_stability_ar(A)
}
\arguments{
\item{A}{s x sk matrix of coefficients as estimated in fit_ar.}
}
\value{
stab_cond list containing the entries 'check' Boolean (TRUE, if stable) and 'poles' (eigenvalues of companion matrix)
}
\description{
check for stability of an AR system (via the eigenvalues of the companion form)
}
