% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{conv_syst}
\alias{conv_syst}
\title{Convert system to canonical form. Input and output are stored in a StSp object.}
\usage{
conv_syst(syst)
}
\arguments{
\item{syst}{a \code{StSp} object}
}
\value{
syst a \code{StSp} object
}
\description{
Convert system to canonical form. Input and output are stored in a StSp object.
}
\examples{
ones_mat <- as.matrix(1,1,1)
syst <- StSp(A = ones_mat*0.5,C=  ones_mat, K = ones_mat, Sigma = ones_mat, type = "innov")
conv_syst(syst)

}
