% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{conv_to_innov}
\alias{conv_to_innov}
\title{Convert system to innovation representation.}
\usage{
conv_to_innov(syst)
}
\arguments{
\item{syst}{a \code{StSp} object}
}
\value{
syst a \code{StSp} object
}
\description{
Convert system to innovation representation.
}
