% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_est.R
\name{cov_seq_grad}
\alias{cov_seq_grad}
\title{Calculate covariances for lags given in 'lags'.}
\usage{
cov_seq_grad(theta, n, s, lags = 0, grad_bool = TRUE)
}
\arguments{
\item{theta}{parameter vector for system.}

\item{n}{integer state dimension}

\item{s}{integer output dimension}

\item{lags}{vector of integers}

\item{grad_bool}{Boolean indicating whether derivatives should be calculated as well.}
}
\value{
matrix of vectorized covariances. First column actual values, remaining columns (if any) derivatives.
}
\description{
Calculate covariances for lags given in 'lags'.
}
