% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cov_Sig_par}
\alias{cov_Sig_par}
\title{Extract Parameters for Estimated Covariance Matrix}
\usage{
cov_Sig_par(Sig)
}
\arguments{
\item{Sig}{matrix representing the estimated covariance matrix of the innovations.}
}
\value{
vector containing the extracted parameters from the Cholesky factor of the covariance matrix.
}
\description{
This function extracts the parameters from the estimated covariance matrix \code{Sig} by performing a Cholesky decomposition
and returns them in vector form. The parameters are extracted from the lower triangular part of the Cholesky factor.
}
\examples{
Sig <- matrix(c(2, 0.5, 0.5, 1), nrow=2, ncol=2)
cov_Sig_par(Sig)

}
