% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{covariance_function}
\alias{covariance_function}
\title{generic method for covariance function}
\usage{
covariance_function(syst, M, plots)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{M}{number of impulses to calculate}

\item{plots}{integer; 0.. no plots; 1.. static plots, 2.. interactive plots}
}
\value{
list of matrices with covariance sequence
}
\description{
Calculate first M covariances for matrix system
Remark: F matrix needs to be stable.
}
