% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{covariance_function.StSp}
\alias{covariance_function.StSp}
\title{Calculate first M covariances for matrix system
F matrix needs to be stable.}
\usage{
\method{covariance_function}{StSp}(syst, M = 20, plots = 0)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{M}{number of impulses to calculate}

\item{plots}{integer; 0.. no plots; 1.. static plots, 2.. interactive plots}
}
\value{
list of matrices with covariance sequence
}
\description{
Calculate first M covariances for matrix system
F matrix needs to be stable.
}
