% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_est.R
\name{dLjapunov}
\alias{dLjapunov}
\title{Calculate the solution to Ljapunov equations (and their derivatives) for a state space system.}
\usage{
dLjapunov(theta, n, s, grad_bool)
}
\arguments{
\item{theta}{vector of parameters}

\item{n}{number of states}

\item{s}{number of outputs}

\item{grad_bool}{if TRUE, also derivative is calculated}
}
\value{
list of matrices. First is solution, the rest the derivatives.
}
\description{
This function calculates the solution to the Ljapunov equation P = F P F' + K Om K' as well as
its derivative with respect to the parameter vector theta (for a system with output dimension s and state dimension n).
The output is given as a list of matrices.
}
