% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_AR_est.R
\name{dLjapunov_AR}
\alias{dLjapunov_AR}
\title{Calculate the solution to Ljapunov equations (and their derivatives) for a
state space system corresponding to the AR system.}
\usage{
dLjapunov_AR(theta, k, s, grad_bool)
}
\arguments{
\item{theta}{vector of parameters}

\item{k}{integer; lag length}

\item{s}{number of outputs}

\item{grad_bool}{if TRUE, also derivative is calculated}
}
\value{
list of matrices. First is solution, the rest the derivatives.
}
\description{
This function calculates the solution to the Ljapunov equation P = F P F' + B Sigma B'
(F being the companion matrix and B=(I,0))
as well as its derivative with respect to the parameter vector theta
(for a system with output dimension s and lag length p).
The output is given as a list of matrices.
}
