% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durbin_levinson.R
\name{durbin_levinson}
\alias{durbin_levinson}
\title{Durbin Levinson algorithm for the estimation of AR models recursive in the lag length.}
\usage{
durbin_levinson(C, L)
}
\arguments{
\item{C}{list of covariances of size sxs.}

\item{L}{integer. maximal lag length}
}
\value{
list of information for estimated AR systems. (I) (a_j(p)), (II) Sigma_p (III) rho_{p,p}.
}
\description{
\code{durbin_levinson} takes covariance sequence (list of covariance matrices) as input as well
as a maximal lag length and calculates sequentially in lag a series of AR models fitted to the
covariances using the Durbin Levinson recursion.
}
\details{
The multivariate version uses two intertwined recursions according to Hannan and Deistler (1988), section 5.7., p. 121. (5.6.13)
(attributed to Whittle).

Each AR models is characterised using (I) the estimated coefficients, (II) the innovation variance Sigma and (III) the PACC.
}
