% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_est.R
\name{est_MD}
\alias{est_MD}
\title{Minimum distance estimation (MD)}
\usage{
est_MD(n, C, lags = 0, W = NULL, theta_init = NULL, print.level = 0)
}
\arguments{
\item{n}{integer representing the order of the system to estimate}

\item{C}{list of estimated covariances}

\item{lags}{integer vector containing the lgas contained in the list C.}

\item{W}{covariance matrix of estimated C.}

\item{theta_init}{initial estimate. If not present, random initialisation is used.}

\item{print.level}{integer representing the level of \code{nlm} output to print. Defaults to \code{0} (print nothing).}
}
\value{
list containing the estimated parameters and the criterion function.
}
\description{
MD estimation for system of order n matches estimated covariances.
}
