% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{est_qml}
\alias{est_qml}
\title{Quasi Maximum Likelihood Estimation (qMLE)}
\usage{
est_qml(y, n, kmax, pe = TRUE, print.level = 0, iterlim = 100)
}
\arguments{
\item{y}{matrix of observed data where rows represent variables and columns represent time points}

\item{n}{integer representing the order of the system to estimate}

\item{kmax}{integer representing the maximum lag to use for the initial autoregressive model fit}

\item{pe}{logical. If TRUE, the prediction error likelihood is used, otherwise the full likelihood is used. Defaults to \code{TRUE}.}

\item{print.level}{integer representing the level of \code{nlm} output to print. Defaults to \code{0} (print nothing).}

\item{iterlim}{integer; maximal number of iterations.}
}
\value{
list containing the estimated parameters; output of nlm..
}
\description{
qML estimation for system of order n using CVA initial estimate
}
