% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UC_utils.R
\name{est_uc}
\alias{est_uc}
\title{Estimate Unobserved Components Model}
\usage{
est_uc(y, theta_init, iterlim = 100, initialize = FALSE, print.level = 0)
}
\arguments{
\item{y}{matrix of observations}

\item{theta_init}{vector of initial parameters}

\item{iterlim}{integer representing the maximum number of iterations for the nonlinear least squares estimation. Defaults to \code{100}.}

\item{initialize}{logical value indicating whether to use an initial estimate for the parameters. Defaults to \code{FALSE}.}

\item{print.level}{integer representing the level of \code{nlm} output to print. Defaults to \code{0}.}
}
\value{
list containing the estimated parameters.
}
\description{
This function estimates the parameters of an unobserved components model using a nonlinear least squares approach.
}
