% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_est.R
\name{find_dim_ident}
\alias{find_dim_ident}
\title{Find the largest state dimension that leads to an identifiable system
given the lags contained in the sequence.}
\usage{
find_dim_ident(s, lags = 0)
}
\arguments{
\item{s}{integer output dimension}

\item{lags}{vector of integers}
}
\value{
min_evs vector of minimal eigenvalues for different state dimensions.
}
\description{
Find the largest state dimension that leads to an identifiable system
given the lags contained in the sequence.
}
