% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{fit_ar}
\alias{fit_ar}
\title{Fit an AR(p) model}
\usage{
fit_ar(y, k = -1, kmax = 10)
}
\arguments{
\item{y}{time series data frame (data contained in columns)}

\item{k}{integer. Lag length. If negative, AIC is used to pick the lag length.}

\item{kmax}{integer. For AIC pick the largest possible lag.}
}
\value{
\code{ar_mod} object, with elements
\describe{
\item{\code{A}}{matrix of AR coefficients.}
\item{\code{AIC}}{AIC values for various lags.}
\item{\code{k}}{optimal lag, if should be determined by AIC, else k like given}
\item{\code{res}}{matrix of residuals}
\item{\code{Sigma}}{estimated variance matrix of residuals}
\item{\code{vcov}}{estimation variance matrix, corresponds to row-wise vectorization}
}
}
\description{
Fit  model with optional AIC selection of optimal lag length
\strong{REMARK}: Subsample is built using kmax lags.
}
