% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{Hankel}
\alias{Hankel}
\title{Calculates the Hankel matrix.}
\usage{
Hankel(syst, krow = 10, kcol = 10)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{krow}{integer; number of block rows}

\item{kcol}{integer; number of block columns.}
}
\value{
reduced order system
}
\description{
Calculates the Hankel matrix.
}
