% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{impulse}
\alias{impulse}
\title{generic method for plotting the impulse response function}
\usage{
impulse(syst, M, plots)
}
\arguments{
\item{syst}{StSp object}

\item{M}{integer; number of lags to include}

\item{plots}{integer; 0... no plots, 1... static plots, 2... interactive plot.}
}
\value{
on screen information
}
\description{
generic method for plotting the impulse response function
}
