% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{impulse.StSp}
\alias{impulse.StSp}
\title{Calculate impulse response for a system
System must be defined as a StSp object.}
\usage{
\method{impulse}{StSp}(syst, M = 20, plots = 0)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{M}{number of impulses to calculate}

\item{plots}{integer; 0.. no plots; 1.. static plots, 2.. interactive plots}
}
\value{
list of matrices with impulse response
}
\description{
Calculate impulse response for a system
System must be defined as a StSp object.
}
