% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{kf}
\alias{kf}
\title{Kalman filter}
\usage{
kf(y, syst, pe = TRUE, init = 0)
}
\arguments{
\item{y}{matrix of observed data where rows represent variables and columns represent time points}

\item{syst}{a \code{StSp} object.}

\item{pe}{logical. If TRUE the prediction error formulation is used. Default is \code{TRUE}.}

\item{init}{numeric. If =0 stationary distribution is assumed, if >0 initial state error variance
equals init*I. Default is \code{0}.}
}
\value{
list xhat, res, sigmat: list of matrices Sigmat, Kt: list of Kalman gains, ll: -2 times
log likelihood.
}
\description{
This function calculates the Kalman filter for system (A, K, C) given as matrices.
General form using Sigma (innovation variance), Q (variance of state noise), S (cross covariance).
}
