% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{lin_filter.StSp}
\alias{lin_filter.StSp}
\title{generic method for filtering}
\usage{
\method{lin_filter}{StSp}(syst, u, xinit = NULL, factor = 1, plots = 0)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{u}{time series to be filtered}

\item{xinit}{initial state (if supplied; default zero initialisation)}

\item{factor}{double; scaling the output}

\item{plots}{integer; 0.. no plots; 1.. static plots, 2.. interactive plots}
}
\value{
filtered series in tsibble form.
}
\description{
Filter using a state space system
}
