% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{mlag}
\alias{mlag}
\title{Produce matrix containing \code{k} lags of matrix \code{y} stapled in block rows.}
\usage{
mlag(y, k)
}
\arguments{
\item{y}{matrix. Each column is one time period.}

\item{k}{integer. Lag length}
}
\value{
matrix. Matrix with \code{k} lags of \code{y}.
}
\description{
\code{mlag} takes a matrix \code{y}, an integer \code{k} and returns a matrix,
where each column contains the \code{k} lags of the corresponding column in \code{y}.
First row of returned matrix start with column \code{k} of \code{y}, last row starts with column 1
}
