% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{model_reduce}
\alias{model_reduce}
\title{Reduces the model to a minimal system eliminating non-obervability and non-controllabilty.
REMARK: Only works for systems in innovation form.}
\usage{
model_reduce(syst, tol = 1e-07)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{tol}{cut of threshhold: singular values below this threshhold are neglected. Default set to 0.0000001.}
}
\value{
reduced order system
}
\description{
Reduces the model to a minimal system eliminating non-obervability and non-controllabilty.
REMARK: Only works for systems in innovation form.
}
