% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{param_syst}
\alias{param_syst}
\title{Obtain system from parameter vector}
\usage{
param_syst(theta, n, s, gradient = FALSE)
}
\arguments{
\item{theta}{list. Parameter vector.}

\item{n}{integer. State dimension.}

\item{s}{integer. Output dimension.}

\item{gradient}{logical; indicates whether the supplied vector are the parameters or the standard deviations.}
}
\value{
syst as a \code{StSp} object
}
\description{
For given parameter theta, obtain system, based on parameter vector (F, K, H), in canonical form
For this we need to know the state dimension n and the output dimension s.
}
\details{
Adds also the lower triangular part of the Cholesky factor for the innovation variance sigma to
the parameter vector.
}
