% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_dist_AR_est.R
\name{param_syst_AR}
\alias{param_syst_AR}
\title{Obtain system from parameter vector}
\usage{
param_syst_AR(theta, k, s)
}
\arguments{
\item{theta}{list. Parameter vector.}

\item{k}{integer. Lag length.}

\item{s}{integer. Output dimension.}
}
\value{
syst as a \code{StSp} object
}
\description{
For given parameter theta, obtain system, based on parameter vector (F, K, H), in canonical form
For this we need to know the state dimension n and the output dimension s.
}
\details{
Adds also the lower triangular part of the Cholesky factor for the innovation variance sigma to
the parameter vector.
}
