% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{param_syst_deriv}
\alias{param_syst_deriv}
\title{Obtain derived system from parameter vector}
\usage{
param_syst_deriv(theta, n, s)
}
\arguments{
\item{theta}{list. Parameter vector.}

\item{n}{integer. State dimension.}

\item{s}{integer. Output dimension.}
}
\value{
systs. A list of \code{StSp} objects. First element is the system, the remaining the various derivatives.
}
\description{
For given parameter theta, obtain the derived system, based on parameter vector (F, K, H), in canonical form
For this we need to know the state dimension n and the output dimension s.
}
\details{
Adds also the lower triangular part of the Cholesky factor for the innovation variance sigma to
the parameter vector.
}
