% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UC_utils.R
\name{param_syst_uc}
\alias{param_syst_uc}
\title{unobserved components model: Convert parameters to system matrices}
\usage{
param_syst_uc(theta)
}
\arguments{
\item{theta}{vector of parameters: C.. 7 params, A: 1 param, Sigma: 1 param. Rest fixed.}
}
\value{
list of system matrices A, K, C, Sigma, Q
}
\description{
This function converts a parameter vector \code{theta} into system matrices for an unobserved components model
for a univariate quarterly time series with a doubly integrated trend, seasonal common cycles and a stationary
added univariate signal.
}
