% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cov.R
\name{plot_cov}
\alias{plot_cov}
\title{plot_cov plots the covariance sequence for a number of systems}
\usage{
plot_cov(syst_vec, channel_in = 1, channel_out = 1, L = 20)
}
\arguments{
\item{syst_vec}{vector of systems or StSp object, if only one system}

\item{channel_in}{integer: SSM input element to plot. Default is \code{1}.}

\item{channel_out}{integer: SSM output element to plot. Default is \code{1}.}

\item{L}{integer, number of lags to include.}
}
\value{
the tibble data file
}
\description{
plot_cov plots the covariance sequence for a number of systems
}
