% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inter.r
\name{plot_inter}
\alias{plot_inter}
\title{Interface for plotting of impulse responses and covariance sequences.}
\usage{
plot_inter(data = list(), type = "IR", interactive = FALSE)
}
\arguments{
\item{data}{list of matrices containing the data}

\item{type}{string; either "IR" for impulse response or "CV" for covariance sequence}

\item{interactive}{Boolean, indicating, if interactive plot is wanted. Default: FALSE.}
}
\value{
x    invisibly returns the input \code{x} after being coerced to an array.
}
\description{
Interface for plotting of impulse responses and covariance sequences.
}
