% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_impulse.R
\name{plot_IR}
\alias{plot_IR}
\title{plot_IR plots the Impulse Response for a number of systems}
\usage{
plot_IR(syst_vec, channel_in = 1, channel_out = 1, L = 20)
}
\arguments{
\item{syst_vec}{vector of systems or StSp object, if only one system}

\item{channel_in}{integer: SSM input element of the impulse response to plot. Default is \code{1}.}

\item{channel_out}{integer: SSM output element of the impulse response to plot. Default is \code{1}.}

\item{L}{integer, number of lags to include.}
}
\value{
the \code{ggplot2} object of the plot
}
\description{
plot_IR plots the Impulse Response for a number of systems
}
