% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mat_lags.R
\name{plot_IR_tibble}
\alias{plot_IR_tibble}
\title{Plot impulse responses and covariance sequences from a collection of systems}
\usage{
plot_IR_tibble(x, inp = 1, outp = 1, main = "", ylab = "")
}
\arguments{
\item{x}{tibble containing the data}

\item{inp}{vector of integers containing the numbers of the input channel}

\item{outp}{vector of integers containing the numbers of the output channel}

\item{main}{title string}

\item{ylab}{y-axis label string}
}
\value{
ggplot2 object
}
\description{
This function plots impulse responses nd covariance sequences for a number of models
selecting the input/output channels to include.
The data is assumed to be stored in a tibble containing the columns lag,in_chan,out_chan,syst.
}
