% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.StSp.R
\name{plot_predict}
\alias{plot_predict}
\title{plot predictions}
\usage{
plot_predict(pr, window = NULL)
}
\arguments{
\item{pr}{list containing the predictions as 'resh' and prediction variances as 'Sig_th'.}

\item{window}{pair of integers specifying the window to plot.}
}
\value{
ggplot2 object containing the plot
}
\description{
plots the predictions
}
