% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.StSp.R
\name{predict.StSp}
\alias{predict.StSp}
\title{prediction function
calculates the predictions}
\usage{
\method{predict}{StSp}(syst, y, h = 1, T_est_val = 1)
}
\arguments{
\item{syst}{StSp system object used for predictions.}

\item{y}{time series data}

\item{h}{prediction horizon}

\item{T_est_val}{time point for separating estimation and validation data}
}
\value{
predictions
}
\description{
prediction function
calculates the predictions
}
