% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{print.StSp}
\alias{print.StSp}
\title{print function: prints a StSp system object in a nice way.}
\usage{
\method{print}{StSp}(x, ...)
}
\arguments{
\item{x}{StSp object}

\item{...}{additional arguments passed on.}
}
\value{
on screen information
}
\description{
print function: prints a StSp system object in a nice way.
}
