% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{sim_ar}
\alias{sim_ar}
\title{Simulate autoregressive time series via companion form of AR parameters}
\usage{
sim_ar(A, Sigma, T = 100, stationary = TRUE)
}
\arguments{
\item{A}{s x sk matrix of coefficients as estimated in fit_ar.}

\item{Sigma}{sxs covariance matrix of noise. Default: I(s).}

\item{T}{integer; sample size}

\item{stationary}{Boolean indicating whether initialisation should be done with a stationary distribution.}
}
\value{
list.
\describe{
\item{\code{ysim}}{simulated time series observations in df format}
}
}
\description{
Receiving the parameters of an AR(p) process, this function creates the according companion matrix for it, and simulates the system.
}
