% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_syst.R
\name{sim_syst}
\alias{sim_syst}
\title{sim_syst simulates time series data from a state space system in innovation form.}
\usage{
sim_syst(syst, T = 100, stationary = FALSE)
}
\arguments{
\item{syst}{\code{StSp} object, which represents a state space model, either in innovation representation or in general form.}

\item{T}{integer; sample length.}

\item{stationary}{Boolean indicating whether the simulated process should be stationary.}
}
\value{
y real sxT matrix of observations.
}
\description{
Time series data of length T is simulated using Gaussian innovations from the state
space system supplied.
The state space system must be in innovation form.
}
