% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_lyap.R
\name{solve_lyap}
\alias{solve_lyap}
\title{solve the Lyapunov equation P = F P F' + Q.}
\usage{
solve_lyap(A = as.matrix(0.5), Q = as.matrix(1))
}
\arguments{
\item{A}{nxn transition matrix of the state space model.}

\item{Q}{nxn state noise variance}
}
\value{
P nxn stationary state variance.
}
\description{
solve the Lyapunov equation P = F P F' + Q.
}
\examples{
# solution to Lyapunov equation 
P <- solve_lyap(A=as.matrix(0.5), Q=as.matrix(1))

}
