% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{solve_shift}
\alias{solve_shift}
\title{solve_shift provides the estimate of F using the shift invariance structure of cO.}
\usage{
solve_shift(cO, s = 1)
}
\arguments{
\item{cO}{observability matrix}

\item{s}{integer, output dimension}
}
\value{
A estimated matrix A
}
\description{
solve_shift provides the estimate of F using the shift invariance structure of cO.
}
