% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{StSp}
\alias{StSp}
\title{user helper function: supports users to define the StSp object.}
\usage{
StSp(
  A = as.matrix(0.5),
  C = as.matrix(1),
  K = as.matrix(1),
  Sigma = as.matrix(1),
  Q = as.matrix(1),
  R = as.matrix(1),
  S = as.matrix(1),
  type = "innov"
)
}
\arguments{
\item{A}{nxn matrix}

\item{C}{sxn matrix}

\item{K}{nxs matrix}

\item{Sigma}{sxs innovation variance matrix}

\item{Q}{nxn matrix of noise variance in state equation}

\item{R}{sxs matrix of noise covariance in observaiton equation}

\item{S}{nxs matrix of cross covariance}

\item{type}{string; default: "innov" for innovation representation. Alternative "UC§ for unobserved components.}
}
\value{
syst StSp object
}
\description{
user helper function: supports users to define the StSp object.
}
