% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar.R
\name{summary.ar_mod}
\alias{summary.ar_mod}
\title{Summary function provides a summary for an estimated AR(k) model.}
\usage{
\method{summary}{ar_mod}(object, ...)
}
\arguments{
\item{object}{... AR(k) model.}

\item{...}{additional arguments passed on.
#'}
}
\value{
output on screen
}
\description{
Summary function provides a summary for an estimated AR(k) model.
}
