% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{summary_qmle}
\alias{summary_qmle}
\title{Present Quasi Maximum Likelihood Estimator (qMLE)}
\usage{
summary_qmle(theta_hat, n = 1, s = 1)
}
\arguments{
\item{theta_hat}{list containing the estimated parameters; output of nlm.}

\item{n}{integer, system dimension}

\item{s}{integer; output dimension}
}
\value{
output on screen.
}
\description{
summary of qMLE output
}
