% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kf_qmle.R
\name{syst_param}
\alias{syst_param}
\title{Extract parameters from system matrices}
\usage{
syst_param(syst)
}
\arguments{
\item{syst}{a \code{StSp} object}
}
\value{
vector containing the extracted F, H, K parameters from the system matrices.
}
\description{
This function extracts the parameters from the system matrices (F, H, K) and returns them in vector form.
}
