% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{Toeplitz}
\alias{Toeplitz}
\title{Calculates the Toeplitz matrix.}
\usage{
Toeplitz(Cv, krow = 10)
}
\arguments{
\item{Cv}{list of covariance matrices}

\item{krow}{integer; number of block rows}
}
\value{
matrix Gamma_p.
}
\description{
Calculates the Toeplitz matrix.
}
