% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{transfer_function}
\alias{transfer_function}
\title{generic method for transfer function (including plotting)}
\usage{
transfer_function(syst, M, plots)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{M}{integer; number of frequencies}

\item{plots}{integer; 0.. no plots; 1.. static plots, 2.. interactive plots}
}
\value{
list of matrices with transfer function at Fourier frequencies
}
\description{
generic method for transfer function (including plotting)
}
