% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{transfer_function.StSp}
\alias{transfer_function.StSp}
\title{generic method for transfer function (including plotting)}
\usage{
\method{transfer_function}{StSp}(syst, M = 100, plots = 0)
}
\arguments{
\item{syst}{system matrices in StSp object form.}

\item{M}{integer; number of points where to evaluate transfer function (default: M=100)}

\item{plots}{integer; 0.. no plots; 1.. static plots, 2.. interactive plots}
}
\value{
list of matrices with transfer function at Fourier frequencies
}
\description{
generic method for transfer function (including plotting)
}
