% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_df.R
\name{tsibble_to_mat}
\alias{tsibble_to_mat}
\title{Convert a \code{\link[tsibble:tsibble]{tsibble}} object to a \code{\link{matrix}}}
\usage{
tsibble_to_mat(ts_obj, var_name = "y")
}
\arguments{
\item{ts_obj}{the \code{\link[tsibble:tsibble]{tsibble::tsibble}} object}

\item{var_name}{string indicating the column of ts_obj containing variable names}
}
\value{
matrix; containing the data in matrix form.
}
\description{
The matrix returned contains the variables in columns, while the time column makes up
the rownames.
}
