% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StSp_cl.R
\name{validate_StSp}
\alias{validate_StSp}
\title{validator method for verifying the created StSp object is valid.}
\usage{
validate_StSp(syst)
}
\arguments{
\item{syst}{StSp object}
}
\value{
syst StSp object
}
\description{
validator method for verifying the created StSp object is valid.
}
